+++ core/src/components.d.ts
@@ -1198,6 +1198,10 @@ export namespace Components {
     */
     'placeholder'?: string | null;
     /**
+    * If `true`, the datetime appears normal but is not interactive.
+    */
+    'readonly': boolean;
+    /**
     * The value of the datetime as a valid ISO 8601 datetime string.
     */
     'value'?: string | null;
@@ -1299,6 +1303,10 @@ export namespace Components {
     * The text to display when there's no date selected yet. Using lowercase to match the input attribute
     */
     'placeholder'?: string | null;
+    /**
+    * If `true`, the datetime appears normal but is not interactive.
+    */
+    'readonly'?: boolean;
     /**
     * The value of the datetime as a valid ISO 8601 datetime string.
     */
@@ -1643,6 +1651,10 @@ export namespace Components {
     * If `true`, the user cannot interact with the input.
     */
     'disabled': boolean;
+    /**
+    * Returns the native `<input>` element used under the hood.
+    */
+    'getInputElement': () => Promise<HTMLInputElement>;
     /**
     * A hint to the browser for which keyboard to display. This attribute applies when the value of the type attribute is `"text"`, `"password"`, `"email"`, or `"url"`. Possible values are: `"verbatim"`, `"latin"`, `"latin-name"`, `"latin-prose"`, `"full-width-latin"`, `"kana"`, `"katakana"`, `"numeric"`, `"tel"`, `"email"`, `"url"`.
     */
@@ -2772,7 +2784,7 @@ export namespace Components {
     */
     'onIonNavDidChange'?: (event: CustomEvent<void>) => void;
     /**
-    * Event fired when the nav will components
+    * Event fired when the nav will change components
     */
     'onIonNavWillChange'?: (event: CustomEvent<void>) => void;
     /**
@@ -3305,6 +3317,10 @@ export namespace Components {
     */
     'name': string;
     /**
+    * The neutral point of the range slider. Default: value is `0` or the `min` when `neutralPoint < min` or `max` when `max < neutralPoint`.
+    */
+    'neutralPoint': number;
+    /**
     * If `true`, a pin with integer value is shown when the knob is pressed.
     */
     'pin': boolean;
@@ -3319,7 +3335,7 @@ export namespace Components {
     /**
     * the value of the range.
     */
-    'value': RangeValue;
+    'value': RangeValue | null;
   }
   interface IonRangeAttributes extends StencilHTMLAttributes {
     /**
@@ -3355,6 +3371,10 @@ export namespace Components {
     */
     'name'?: string;
     /**
+    * The neutral point of the range slider. Default: value is `0` or the `min` when `neutralPoint < min` or `max` when `max < neutralPoint`.
+    */
+    'neutralPoint'?: number;
+    /**
     * Emitted when the range loses focus.
     */
     'onIonBlur'?: (event: CustomEvent<void>) => void;
@@ -3381,7 +3401,7 @@ export namespace Components {
     /**
     * the value of the range.
     */
-    'value'?: RangeValue;
+    'value'?: RangeValue | null;
   }
 
   interface IonRefresherContent {
@@ -3506,7 +3526,7 @@ export namespace Components {
     */
     'disabled'?: boolean;
     /**
-    * Event that needs to be listen to in order to respond to reorder action. `ion-reorder-group` uses this event to delegate to the user the reordering of data array.   The complete() method exposed as
+    * Event that needs to be listened to in order to complete the reorder action. Once the event has been emitted, the `complete()` method then needs to be called in order to finalize the reorder action.
     */
     'onIonItemReorder'?: (event: CustomEvent<ItemReorderEventDetail>) => void;
   }
@@ -3620,7 +3640,7 @@ export namespace Components {
     /**
     * Go back to previous page in the window.history.
     */
-    'goBack': () => Promise<void>;
+    'back': () => Promise<void>;
     'navChanged': (direction: RouterDirection) => Promise<boolean>;
     'printDebug': () => void;
     /**
@@ -3691,6 +3711,10 @@ export namespace Components {
     * Set the amount of time, in milliseconds, to wait to trigger the `ionChange` event after each keystroke.
     */
     'debounce': number;
+    /**
+    * Returns the native `<input>` element used under the hood.
+    */
+    'getInputElement': () => Promise<HTMLInputElement>;
     /**
     * The mode determines which platform styles to use.
     */
@@ -4181,6 +4205,10 @@ export namespace Components {
     * Update the underlying slider implementation. Call this if you've added or removed child slides.
     */
     'update': () => Promise<void>;
+    /**
+    * Force swiper to update its height (when autoHeight enabled) for the duration equal to 'speed' parameter
+    */
+    'updateAutoHeight': (speed?: number | undefined) => Promise<void>;
   }
   interface IonSlidesAttributes extends StencilHTMLAttributes {
     /**
@@ -4530,6 +4558,10 @@ export namespace Components {
     * If `true`, the user cannot interact with the textarea.
     */
     'disabled': boolean;
+    /**
+    * Returns the native `<textarea>` element used under the hood.
+    */
+    'getInputElement': () => Promise<HTMLTextAreaElement>;
     /**
     * If the value of the type attribute is `text`, `email`, `search`, `password`, `tel`, or `url`, this attribute specifies the maximum number of characters that the user can enter.
     */