+++ core/src/components/button/button.tsx
@@ -4,6 +4,12 @@ import { Color, Mode, RouterDirection } from '../../interface';
 import { hasShadowDom } from '../../utils/helpers';
 import { createColorClasses, openURL } from '../../utils/theme';
 
+/**
+ * @slot - Content is placed between the named slots if provided without a slot.
+ * @slot icon-only - Should be used on an icon in a button that has no text.
+ * @slot start - Content is placed to the left of the button text in LTR, and to the right in RTL.
+ * @slot end - Content is placed to the right of the button text in LTR, and to the left in RTL.
+ */
 @Component({
   tag: 'ion-button',
   styleUrls: {
@@ -104,7 +110,7 @@ export class Button implements ComponentInterface {
   @Listen('click')
   onClick(ev: Event) {
     if (this.type === 'button') {
-      return openURL(this.win, this.href, ev, this.routerDirection);
+      openURL(this.win, this.href, ev, this.routerDirection);
 
     } else if (hasShadowDom(this.el)) {
       // this button wants to specifically submit a form
@@ -122,7 +128,6 @@ export class Button implements ComponentInterface {
         fakeButton.remove();
       }
     }
-    return Promise.resolve(false);
   }
 
   private onFocus = () => {