+++ core/src/components/range/range.tsx
@@ -1,9 +1,13 @@
-import { Component, ComponentInterface, Element, Event, EventEmitter, Prop, QueueApi, State, Watch } from '@stencil/core';
+import { Component, ComponentInterface, Element, Event, EventEmitter, Listen, Prop, QueueApi, State, Watch } from '@stencil/core';
 
 import { Color, Gesture, GestureDetail, KnobName, Mode, RangeChangeEventDetail, RangeValue, StyleEventDetail } from '../../interface';
 import { clamp, debounceEvent } from '../../utils/helpers';
 import { createColorClasses, hostContext } from '../../utils/theme';
 
+/**
+ * @slot start - Content is placed to the left of the range slider in LTR, and to the right in RTL.
+ * @slot end - Content is placed to the right of the range slider in LTR, and to the left in RTL.
+ */
 @Component({
   tag: 'ion-range',
   styleUrls: {
@@ -39,6 +43,25 @@ export class Range implements ComponentInterface {
    * The mode determines which platform styles to use.
    */
   @Prop() mode!: Mode;
+
+  /**
+   * The neutral point of the range slider.
+   * Default: value is `0` or the `min` when `neutralPoint < min` or `max` when `max < neutralPoint`.
+   */
+  @Prop() neutralPoint = 0;
+  protected neutralPointChanged() {
+    if (this.noUpdate) {
+      return;
+    }
+    const { min, max, neutralPoint } = this;
+
+    if (max < neutralPoint) {
+      this.neutralPoint = max;
+    }
+    if (neutralPoint < min) {
+      this.neutralPoint = min;
+    }
+  }
 
   /**
    * How long, in milliseconds, to wait to trigger the
@@ -115,7 +138,7 @@ export class Range implements ComponentInterface {
   /**
    * the value of the range.
    */
-  @Prop({ mutable: true }) value: RangeValue = 0;
+  @Prop({ mutable: true }) value: RangeValue | null = null;
   @Watch('value')
   protected valueChanged(value: RangeValue) {
     if (!this.noUpdate) {
@@ -145,6 +168,24 @@ export class Range implements ComponentInterface {
    */
   @Event() ionBlur!: EventEmitter<void>;
 
+  @Listen('focusout')
+  onBlur() {
+    if (this.hasFocus) {
+      this.hasFocus = false;
+      this.ionBlur.emit();
+      this.emitStyle();
+    }
+  }
+
+  @Listen('focusin')
+  onFocus() {
+    if (!this.hasFocus) {
+      this.hasFocus = true;
+      this.ionFocus.emit();
+      this.emitStyle();
+    }
+  }
+
   componentWillLoad() {
     this.updateRatio();
     this.debounceChanged();
@@ -165,7 +206,14 @@ export class Range implements ComponentInterface {
     this.gesture.setDisabled(this.disabled);
   }
 
-  private handleKeyboard = (knob: string, isIncrease: boolean) => {
+  componentDidUnload() {
+    if (this.gesture) {
+      this.gesture.destroy();
+      this.gesture = undefined;
+    }
+  }
+
+  private handleKeyboard = (knob: KnobName, isIncrease: boolean) => {
     let step = this.step;
     step = step > 0 ? step : 1;
     step = step / (this.max - this.min);
@@ -181,14 +229,14 @@ export class Range implements ComponentInterface {
   }
 
   private getValue(): RangeValue {
-    const value = this.value || 0;
+    const value = this.value || this.neutralPoint || 0;
     if (this.dualKnobs) {
       if (typeof value === 'object') {
         return value;
       }
       return {
-        lower: 0,
-        upper: value
+        lower: this.value === null ? this.neutralPoint : 0,
+        upper: this.value === null ? this.neutralPoint : value
       };
     } else {
       if (typeof value === 'object') {
@@ -200,39 +248,28 @@ export class Range implements ComponentInterface {
 
   private emitStyle() {
     this.ionStyle.emit({
+      'interactive': true,
       'interactive-disabled': this.disabled
     });
   }
 
-  private fireBlur() {
-    if (this.hasFocus) {
-      this.hasFocus = false;
-      this.ionBlur.emit();
-      this.emitStyle();
-    }
-  }
-
-  private fireFocus() {
-    if (!this.hasFocus) {
-      this.hasFocus = true;
-      this.ionFocus.emit();
-      this.emitStyle();
-    }
-  }
-
   private onStart(detail: GestureDetail) {
-    this.fireFocus();
-
     const rect = this.rect = this.rangeSlider!.getBoundingClientRect() as any;
     const currentX = detail.currentX;
 
     // figure out which knob they started closer to
-    const ratio = clamp(0, (currentX - rect.left) / rect.width, 1);
+    let ratio = clamp(0, (currentX - rect.left) / rect.width, 1);
+    if (document.dir === 'rtl') {
+      ratio = 1 - ratio;
+    }
+
     this.pressedKnob =
       !this.dualKnobs ||
       Math.abs(this.ratioA - ratio) < Math.abs(this.ratioB - ratio)
         ? 'A'
         : 'B';
+
+    this.setFocus(this.pressedKnob);
 
     // update the active knob's position
     this.update(currentX);
@@ -245,7 +282,6 @@ export class Range implements ComponentInterface {
   private onEnd(detail: GestureDetail) {
     this.update(detail.currentX);
     this.pressedKnob = undefined;
-    this.fireBlur();
   }
 
   private update(currentX: number) {
@@ -253,10 +289,17 @@ export class Range implements ComponentInterface {
     // update the knob being interacted with
     const rect = this.rect;
     let ratio = clamp(0, (currentX - rect.left) / rect.width, 1);
+    if (document.dir === 'rtl') {
+      ratio = 1 - ratio;
+    }
+
     if (this.snaps) {
       // snaps the ratio to the current value
-      const value = ratioToValue(ratio, this.min, this.max, this.step);
-      ratio = valueToRatio(value, this.min, this.max);
+      ratio = valueToRatio(
+        ratioToValue(ratio, this.min, this.max, this.step),
+        this.min,
+        this.max
+      );
     }
 
     // update which knob is pressed
@@ -317,6 +360,15 @@ export class Range implements ComponentInterface {
     this.noUpdate = false;
   }
 
+  private setFocus(knob: KnobName) {
+    if (this.el.shadowRoot) {
+      const knobEl = this.el.shadowRoot.querySelector(knob === 'A' ? '.range-knob-a' : '.range-knob-b') as HTMLElement | undefined;
+      if (knobEl) {
+        knobEl.focus();
+      }
+    }
+  }
+
   hostData() {
     return {
       class: {
@@ -328,35 +380,101 @@ export class Range implements ComponentInterface {
       }
     };
   }
+  protected getActiveBarPosition() {
+    const { min, max, neutralPoint, ratioLower, ratioUpper } = this;
+    const neutralPointRatio = valueToRatio(neutralPoint, min, max);
+
+    // dual knob handling
+    let left = `${ratioLower * 100}%`;
+    let right = `${100 - ratioUpper * 100}%`;
+
+    // single knob handling
+    if (!this.dualKnobs) {
+      if (this.ratioA < neutralPointRatio) {
+        right = `${neutralPointRatio * 100}%`;
+        left = `${this.ratioA * 100}%`;
+      } else {
+        right = `${100 - this.ratioA * 100}%`;
+        left = `${neutralPointRatio * 100}%`;
+      }
+    }
+
+    return {
+      left,
+      right
+    };
+  }
+
+  protected isTickActive(stepRatio: number) {
+    const { min, max, neutralPoint, ratioLower, ratioUpper } = this;
+    const neutralPointRatio = valueToRatio(neutralPoint, min, max);
+
+    if (this.dualKnobs) {
+      return (stepRatio >= ratioLower && stepRatio <= ratioUpper);
+    }
+
+    if (this.ratioA <= neutralPointRatio && stepRatio >= this.ratioA && stepRatio <= neutralPointRatio) {
+      return true;
+    }
+
+    if (this.ratioA >= neutralPointRatio && stepRatio <= this.ratioA && stepRatio >= neutralPointRatio) {
+      return true;
+    }
+
+    return false;
+  }
 
   render() {
-    const { min, max, step, ratioLower, ratioUpper } = this;
+    const { min, max, neutralPoint, step } = this;
+    const barPosition = this.getActiveBarPosition();
 
-    const barL = `${ratioLower * 100}%`;
-    const barR = `${100 - ratioUpper * 100}%`;
+    const isRTL = document.dir === 'rtl';
+    const start = isRTL ? 'right' : 'left';
+    const end = isRTL ? 'left' : 'right';
 
-    const ticks = [];
+    const ticks: any[] = [];
+
     if (this.snaps) {
       for (let value = min; value <= max; value += step) {
         const ratio = valueToRatio(value, min, max);
-        ticks.push({
+
+        const tick: any = {
           ratio,
-          active: ratio >= ratioLower && ratio <= ratioUpper,
-          left: `${ratio * 100}%`
-        });
+          active: this.isTickActive(ratio),
+        };
+
+        tick[start] = `${ratio * 100}%`;
+
+        ticks.push(tick);
       }
     }
+
+    const tickStyle = (tick: any) => {
+      const style: any = {};
+      style[start] = tick[start];
+
+      return style;
+    };
+
+    const barStyle = () => {
+      const style: any = {};
+
+      style[start] = barPosition[start];
+      style[end] = barPosition[end];
+
+      return style;
+    };
 
     return [
       <slot name="start"></slot>,
       <div class="range-slider" ref={el => this.rangeSlider = el}>
-        {ticks.map(t => (
+        {ticks.map(tick => (
           <div
-            style={{ left: t.left }}
+            style={tickStyle(tick)}
             role="presentation"
             class={{
               'range-tick': true,
-              'range-tick-active': t.active
+              'range-tick-active': tick.active
             }}
           />
         ))}
@@ -365,10 +483,7 @@ export class Range implements ComponentInterface {
         <div
           class="range-bar range-bar-active"
           role="presentation"
-          style={{
-            left: barL,
-            right: barR
-          }}
+          style={barStyle()}
         />
 
         { renderKnob({
@@ -380,7 +495,8 @@ export class Range implements ComponentInterface {
           disabled: this.disabled,
           handleKeyboard: this.handleKeyboard,
           min,
-          max
+          max,
+          neutralPoint
         })}
 
         { this.dualKnobs && renderKnob({
@@ -392,7 +508,8 @@ export class Range implements ComponentInterface {
           disabled: this.disabled,
           handleKeyboard: this.handleKeyboard,
           min,
-          max
+          max,
+          neutralPoint
         })}
       </div>,
       <slot name="end"></slot>
@@ -401,19 +518,31 @@ export class Range implements ComponentInterface {
 }
 
 interface RangeKnob {
-  knob: string;
+  knob: KnobName;
   value: number;
   ratio: number;
   min: number;
   max: number;
+  neutralPoint: number;
   disabled: boolean;
   pressed: boolean;
   pin: boolean;
 
-  handleKeyboard: (name: string, isIncrease: boolean) => void;
+  handleKeyboard: (name: KnobName, isIncrease: boolean) => void;
 }
 
-function renderKnob({ knob, value, ratio, min, max, disabled, pressed, pin, handleKeyboard }: RangeKnob) {
+function renderKnob({ knob, value, ratio, min, max, neutralPoint, disabled, pressed, pin, handleKeyboard }: RangeKnob) {
+  const isRTL = document.dir === 'rtl';
+  const start = isRTL ? 'right' : 'left';
+
+  const knobStyle = () => {
+    const style: any = {};
+
+    style[start] = `${ratio * 100}%`;
+
+    return style;
+  };
+
   return (
     <div
       onKeyDown={(ev: KeyboardEvent) => {
@@ -431,19 +560,21 @@ function renderKnob({ knob, value, ratio, min, max, disabled, pressed, pin, hand
       }}
       class={{
         'range-knob-handle': true,
+        'range-knob-a': knob === 'A',
+        'range-knob-b': knob === 'B',
         'range-knob-pressed': pressed,
         'range-knob-min': value === min,
-        'range-knob-max': value === max
+        'range-knob-max': value === max,
+        'range-knob-neutral': value === neutralPoint
       }}
-      style={{
-        'left': `${ratio * 100}%`
-      }}
+      style={knobStyle()}
       role="slider"
       tabindex={disabled ? -1 : 0}
       aria-valuemin={min}
       aria-valuemax={max}
       aria-disabled={disabled ? 'true' : null}
       aria-valuenow={value}
+      aria-valueneutral={neutralPoint}
     >
       {pin && <div class="range-pin" role="presentation">{Math.round(value)}</div>}
       <div class="range-knob" role="presentation" />